LPS-125131 Move Disposable, EventHandle, EventHandler, and EventEmitter to events folder and update export
