+++ modules/apps/frontend-js/frontend-js-web/src/main/resources/META-INF/resources/index.es.js
@@ -42,9 +42,9 @@ export {default as BREAKPOINTS} from './liferay/breakpoints';
 export {default as CompatibilityEventProxy} from './liferay/CompatibilityEventProxy.es';
 
 export {default as DefaultEventHandler} from './liferay/DefaultEventHandler.es';
-export {default as Disposable} from './liferay/Disposable';
-export {default as EventEmitter} from './liferay/EventEmitter';
-export {default as EventHandler} from './liferay/EventHandler';
+export {default as Disposable} from './liferay/events/Disposable';
+export {default as EventEmitter} from './liferay/events/EventEmitter';
+export {default as EventHandler} from './liferay/events/EventHandler';
 export {default as ItemSelectorDialog} from './liferay/ItemSelectorDialog.es';
 export {default as PortletBase} from './liferay/PortletBase.es';
