+++ modules/apps/frontend-js/frontend-js-web/src/main/resources/META-INF/resources/liferay/side_navigation.es.js
@@ -12,7 +12,7 @@
  * details.
  */
 
-import EventEmitter from './EventEmitter';
+import EventEmitter from './events/EventEmitter';
 
 /**
  * Options
